KlayGE: MsgInput: Fix the compiling error on Win7/8 code path. (Github #139)

+++ KlayGE/Plugins/Src/Input/MsgInput/MISensor.cpp
@@ -316,7 +316,7 @@ namespace KlayGE
 			if (SUCCEEDED(hr))
 			{
 				IID REPORT_TYPES[] = { IID_ILatLongReport };
-				hr = location->RequestPermissions(nullptr, REPORT_TYPES, std::size(REPORT_TYPES), true);
+				hr = location->RequestPermissions(nullptr, REPORT_TYPES, static_cast<ULONG>(std::size(REPORT_TYPES)), true);
 				if (SUCCEEDED(hr))
 				{
 					locator_ = MakeCOMPtr(location);
